chore: Fix container name in deploy workflow command
This commit updates the deploy workflow to fix the container name in the command used to execute commands inside the cream-api container. The previous command used the output of `docker ps -qf "name=cream-api"` to get the container ID, but it should have used the container name directly. This change ensures that the commands are executed correctly inside the cream-api container.